// server-app/src/modules/auth/models/renewTokens.model.ts
import z from 'zod';

const responseSchema = {
  200: z.object({
    statusCode: z.literal(200),
    accessToken: z.string(),
  }),
  '4xx': z.object({
    statusCode: z.number(),
    error: z.string(),
    message: z.string(),
  }),
};

export const signOutSchema = {
  response: responseSchema,
};

export type RenewTokensResponse = z.infer<(typeof responseSchema)[200]>;